+++ pyat/at/lattice/constants.py
@@ -2,6 +2,7 @@ from math import sqrt as _sqrt, pi
 # noinspection PyUnresolvedReferences
 from scipy.constants import c as clight
 from scipy.constants import physical_constants as _cst
+from scipy.constants import e as qe
 
 e_mass = 1.0e+06 * _cst['electron mass energy equivalent in MeV'][0]  # eV
 p_mass = 1.0e+06 * _cst['proton mass energy equivalent in MeV'][0]    # eV
